quinta

+++ src/services/api.js
@@ -14,7 +14,10 @@ api.interceptors.request.use(config => {
 api.interceptors.response.use(
   response => response,
   err => {
-    if (err.response.status === 401) {
+    if (
+      err.response.status === 401 &&
+      err.response.data.error === 'Token Invalid'
+    ) {
       localStorage.clear();
       alert('Seção expirada, refaça o Login');
       window.location.reload();
